requirements.txt
# Core dependencies
discord.py>=2.3.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
pydantic>=2.0.0
typing-extensions>=4.0.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code quality tools
ruff>=0.1.0
black>=23.0.0
mypy>=1.5.0
pre-commit>=3.4.0
safety>=2.3.0
bandit>=1.7.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.0.0
mkdocstrings>=0.23.0

# CI/CD helpers
github3.py>=4.0.0
